bugfix for overlay positions on blank endpoints
Seems to have been a bug with the overlay position for blank endpoints. Since the last array element was 0, the y-position of the overlay for this endpoint was multiplied with 0 and therefore didn't change. Changed the last two array elements to 1,1 instead of 10,0 so the positioning of overlays on blank endpoints is now pixel precise.